STP fix, only try contexts on Cisco Only Cisco is known to support this type of polling
